requirements.txt
asgiref
boto3
botocore
Brotli
cachetools
certifi
cffi
charset-normalizer
cryptography
dj-database-url
Django
django-anymail[sendgrid]
django-allauth
django-environ
django-storages
djangorestframework
google-auth
google-auth-oauthlib
gunicorn
idna
jmespath
mercadopago
oauthlib
packaging
pillow
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
PyJWT
requests
requests-oauthlib
rsa
s3transfer
six
sqlparse
typing_extensions
tzdata
urllib3
whitenoise
